refactor(workflow): merge WorkflowEngine + AppState into WorkflowCtx
Collapses the redundant two-struct indirection into one orchestrator
type that also serves as the axum state (Arc<WorkflowCtx>). Drops the
S generic parameter via Arc<dyn WorkflowStore>. Flattens engine/
submodule dir into top-level files with descriptive names
(lifecycle/children/activities/tasks/signals/workers/schedules/
namespaces).

Zero behaviour change. 1092 tests pass (same as baseline), 1 doctest
ignored (preserved from WorkflowEngine → WorkflowCtx method rename).

New files at crates/assay-workflow/src/:
  ctx.rs            — WorkflowCtx struct + start() + BackgroundTasks
                      wrapper + broadcast helper + module-private
                      helpers (strip_continued_suffix, timestamp_now,
                      inject_engine_version) + engine_version_stamp
                      tests
  lifecycle.rs      — start/get/list/cancel/terminate/complete/fail
                      workflow methods
  children.rs       — start_child, list_children, continue_as_new,
                      notify_parent_of_child_outcome
  activities.rs     — schedule/claim/get/complete/fail/heartbeat
                      + record_side_effect
  tasks.rs          — claim_workflow_task + submit_workflow_commands
                      + schedule_timer
  signals.rs        — send_signal + get_events
  workers.rs        — register/heartbeat/list_worker
  schedules.rs      — schedule CRUD + set_schedule_paused
  namespaces.rs     — namespace CRUD + snapshots
  auth_mode.rs      — AuthMode config moved out of api/mod.rs

Deleted:
  engine/           — all 6 submodules (mod + workflows + children +
                      activities + tasks + collections) absorbed into
                      the top-level files above. AppState<S> removed
                      from api/mod.rs.

Handler signatures across crates/assay-workflow/src/api/*.rs:
  State<Arc<AppState<S>>> → State<Arc<WorkflowCtx>>  (generic gone)
  state.engine.foo(...) → state.foo(...)

Runtime binary (crates/assay/src/main.rs) and tests updated to
WorkflowCtx::start(Arc::new(store)) pattern.

See plan 12a Task 1.3 (revised).
