+++ crates/assay-workflow/src/api/workers.rs
@@ -6,10 +6,10 @@ use axum::{Json, Router};
 use serde::Deserialize;
 
 use crate::api::workflows::AppError;
-use crate::api::AppState;
+use crate::ctx::WorkflowCtx;
 use crate::store::WorkflowStore;
 
-pub fn router<S: WorkflowStore + 'static>() -> Router<Arc<AppState<S>>> {
+pub fn router<S: WorkflowStore + 'static>() -> Router<Arc<WorkflowCtx<S>>> {
     Router::new().route("/workers", get(list_workers))
 }
 
@@ -32,10 +32,10 @@ fn default_namespace() -> String {
     ),
 )]
 pub async fn list_workers<S: WorkflowStore>(
-    State(state): State<Arc<AppState<S>>>,
+    State(state): State<Arc<WorkflowCtx<S>>>,
     Query(q): Query<NsQuery>,
 ) -> Result<Json<Vec<serde_json::Value>>, AppError> {
-    let workers = state.engine.list_workers(&q.namespace).await?;
+    let workers = state.list_workers(&q.namespace).await?;
     let json: Vec<serde_json::Value> = workers
         .into_iter()
         .map(|w| serde_json::to_value(w).unwrap_or_default())
