+++ crates/assay-workflow/tests/api_integration.rs
@@ -1,19 +1,13 @@
-use assay_workflow::{WorkflowEngine, SqliteStore};
+use assay_workflow::{SqliteStore, WorkflowCtx};
 use std::sync::Arc;
-use tokio::sync::broadcast;
 
 /// Helper: start engine + API on a random port, return the base URL.
 async fn start_test_server() -> (String, tokio::task::JoinHandle<()>) {
     let store = SqliteStore::new("sqlite::memory:").await.unwrap();
-    let engine = WorkflowEngine::start(store);
-
-    let (event_tx, _) = broadcast::channel(64);
-    let state = Arc::new(assay_workflow::api::AppState {
-        engine: Arc::new(engine),
-        event_tx,
-        auth_mode: assay_workflow::api::auth::AuthMode::no_auth(),
-        binary_version: None,
-    });
+    let state = Arc::new(
+        WorkflowCtx::start(Arc::new(store))
+            .with_auth_mode(assay_workflow::api::auth::AuthMode::no_auth()),
+    );
 
     let app = assay_workflow::api::router(state);
     let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
